--- a/Blockzilla/AppDelegate.swift
+++ b/Blockzilla/AppDelegate.swift
@@ -148,10 +148,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate, ModalDelegate, AppSplashC
     private func setStateForUITesting() {
         if AppDelegate.isUITestingEnabled {
             // If you need reset your app to clear state
-            // UserDefaults.standard.removePersistentDomain(forName: Bundle.main.bundleIdentifier!)
+            UserDefaults.standard.removePersistentDomain(forName: Bundle.main.bundleIdentifier!)
 
             // To speed up your tests
-            UIApplication.shared.keyWindow?.layer.speed = 200
+            UIApplication.shared.keyWindow?.layer.speed = 300
             UIView.setAnimationsEnabled(false)
         }
     }