This PR comes afer this info was shared https://twitter.com/twannl/status/1407986786299691010
So, trying to use that and compare results to see if it is worthy to add