@isabelrios 2021-07-09
Some data after running the different test plans with and without this solution.
Selecting speed 200. These are the results when:

|  | With Changes  | Without Changes | Time saved
|--| ---------------------------------- |:-------------:|:-------------:|
| Smoke Test Suite (8 tests) | 236 s - 3.9 min  |  246 s - 4.1 | 10 s
| Full Functional Test Suite (31 tests) | 800 s - 13 min | 829 s - 13.8 min | 20 s |

When running the tests looks like the app actually loads and goes faster but the time saved in the end is not that much.  In Bitrise we can see some builds around 5.30 mins with this solution and most of them around 6 and just a few seconds. But not a very big decrease as said in the [article](https://narlei.com/development/how-speed-up-your-ui-tests-on-ios/). 

That said, we can try to use this solution for a while, saving time is always good.

@st3fan happy to hear your feedback here. Thanks!