@st3fan 2021-07-09
No magic here ✨ Not sure if it is worth it then?