@isabelrios 2021-07-09
> No magic here ✨ Not sure if it is worth it then?

not very significant unless I'm missing something.. 🤷‍♀️ currently we are saving ~1 second per test... Not sure either if it is worthy to try.. maybe we can set this for sometime and see if we notice some improvement...