@isabelrios 2021-07-19
Closing this PR for now since we are not going to apply these changes, not improving the build execution time significantly.